+++ contracts/fungible-tokens/andromeda-cw20-redeem/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod contract;
+pub mod execute;
 #[cfg(all(not(target_arch = "wasm32"), feature = "testing"))]
 pub mod mock;
+pub mod query;
 
 mod state;
