+++ contracts/fungible-tokens/andromeda-cw20-redeem/src/mock.rs
@@ -2,9 +2,10 @@
 
 use crate::contract::{execute, instantiate, query};
 use andromeda_fungible_tokens::cw20_redeem::{Cw20HookMsg, ExecuteMsg, InstantiateMsg, QueryMsg};
-use andromeda_std::amp::{AndrAddr, Recipient};
+use andromeda_std::amp::Recipient;
 use andromeda_std::common::{expiration::Expiry, MillisecondsDuration};
 use cosmwasm_std::{Empty, Uint128};
+use cw_asset::AssetInfo;
 use cw_multi_test::{Contract, ContractWrapper};
 
 pub fn mock_andromeda_cw20_redeem() -> Box<dyn Contract<Empty>> {
@@ -13,24 +14,24 @@ pub fn mock_andromeda_cw20_redeem() -> Box<dyn Contract<Empty>> {
 }
 
 pub fn mock_cw20_redeem_instantiate_msg(
-    token_address: String,
     kernel_address: String,
     owner: Option<String>,
 ) -> InstantiateMsg {
     InstantiateMsg {
-        token_address: AndrAddr::from_string(token_address),
         kernel_address,
         owner,
     }
 }
 
 pub fn mock_cw20_redeem_start_redemption_condition_hook_msg(
+    redeemed_asset: AssetInfo,
     exchange_rate: Uint128,
     recipient: Option<Recipient>,
     start_time: Option<Expiry>,
     duration: Option<MillisecondsDuration>,
 ) -> Cw20HookMsg {
     Cw20HookMsg::StartRedemptionCondition {
+        redeemed_asset,
         exchange_rate,
         recipient,
         start_time,
@@ -42,13 +43,19 @@ pub fn mock_cw20_redeem_hook_redeem_msg() -> Cw20HookMsg {
     Cw20HookMsg::Redeem {}
 }
 
+pub fn mock_redeem_msg() -> ExecuteMsg {
+    ExecuteMsg::Redeem {}
+}
+
 pub fn mock_cw20_set_redemption_condition_native_msg(
+    redeemed_asset: AssetInfo,
     exchange_rate: Uint128,
     recipient: Option<Recipient>,
     start_time: Option<Expiry>,
     duration: Option<MillisecondsDuration>,
 ) -> ExecuteMsg {
     ExecuteMsg::SetRedemptionCondition {
+        redeemed_asset,
         exchange_rate,
         recipient,
         start_time,
